+++ src/hooks/useSignInMutation.ts
@@ -1,0 +1,42 @@
+import { useMutation } from "@tanstack/react-query";
+import { useNavigate } from "react-router-dom";
+import { loginRequestDto } from "@/api/member/type";
+import { postLogin } from "@/api/member";
+import { saveAccessToken, saveRefreshToken } from "@/utils/manageTokenInfo";
+import { MODALS } from "@/constants/modals";
+import { useState } from "react";
+import { PAGE_PATH } from "@/constants/pagePath";
+
+export const useSignInMutation = () => {
+  const navigate = useNavigate();
+  const [currentModal, setCurrentModal] = useState<string | null>(null);
+
+  const { mutateAsync: mutateAsyncForToken } = useMutation({
+    mutationFn: (enteredSignInInfo: loginRequestDto) => postLogin(enteredSignInInfo),
+    onError: (error: any) => {
+      const statusCode = error?.response?.status;
+      switch (statusCode) {
+        case 401:
+          setCurrentModal(MODALS.invalidPassword.id);
+          break;
+        case 404:
+          setCurrentModal(MODALS.invalidEmail.id);
+          break;
+        default:
+          alert("로그인 처리 중 오류가 발생했습니다. 다시 시도해주세요.");
+      }
+    },
+  });
+
+  const postSignIn = async (enteredSignInInfo: loginRequestDto) => {
+    const tokens = await mutateAsyncForToken(enteredSignInInfo);
+    if (tokens?.accessToken) {
+      saveAccessToken(tokens.accessToken);
+      saveRefreshToken(tokens.refreshToken);
+      alert("로그인 성공!");
+      navigate(PAGE_PATH.home);
+    }
+  };
+
+  return { postSignIn, currentModal, setCurrentModal };
+};
